move robot command and state classes into a helper file to avoid circular imports in modules
